# admin swagger.json
{
   "swagger":"2.0",
   "info":{
      "description":"Api Documentation",
      "version":"1.0",
      "title":"Api Documentation",
      "termsOfService":"urn:tos",
      "contact":{

      },
      "license":{
         "name":"Apache 2.0",
         "url":"http://www.apache.org/licenses/LICENSE-2.0"
      }
   },
   "host":"localhost:8282",
   "basePath":"/",
...
   "paths":{
...
      "/customer":{
         "get":{
            "tags":[
               "admin-controller"
            ],
            "summary":"getAllCustomers",
            "operationId":"getAllCustomersUsingGET",
            "produces":[
               "*/*"
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/Customer"
                     }
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "post":{
            "tags":[
               "admin-controller"
            ],
            "summary":"saveCustomer",
            "operationId":"saveCustomerUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"o",
                  "description":"o",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Customer"
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Created",
                  "schema":{
                     "$ref":"#/definitions/Customer"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "put":{
            "tags":[
               "admin-controller"
            ],
            "summary":"updateCustomer",
            "operationId":"updateCustomerUsingPUT",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"o",
                  "description":"o",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Customer"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK"
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/customer/{id}":{
         "get":{
            "tags":[
               "admin-controller"
            ],
            "summary":"getCustomer",
            "operationId":"getCustomerUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/Customer"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "delete":{
            "tags":[
               "admin-controller"
            ],
            "summary":"deleteCustomer",
            "operationId":"deleteCustomerUsingDELETE",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{
               "204":{
                  "description":"No Content"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               }
            },
            "deprecated":false
         }
      },
...
      "/invoice":{
         "get":{
            "tags":[
               "admin-controller"
            ],
            "summary":"getAllInvoices",
            "operationId":"getAllInvoicesUsingGET",
            "produces":[
               "*/*"
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/Invoice"
                     }
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "post":{
            "tags":[
               "admin-controller"
            ],
            "summary":"submitInvoice",
            "operationId":"submitInvoiceUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"invoice",
                  "description":"invoice",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Invoice"
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Created",
                  "schema":{
                     "$ref":"#/definitions/Invoice"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "put":{
            "tags":[
               "admin-controller"
            ],
            "summary":"updateLevelUp",
            "operationId":"updateLevelUpUsingPUT",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"invoice",
                  "description":"invoice",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Invoice"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK"
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/invoice/customer/{id}":{
         "get":{
            "tags":[
               "admin-controller"
            ],
            "summary":"getInvoicesByCustomerId",
            "operationId":"getInvoicesByCustomerIdUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/Invoice"
                     }
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/invoice/{id}":{
         "get":{
            "tags":[
               "admin-controller"
            ],
            "summary":"getInvoiceById",
            "operationId":"getInvoiceByIdUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/Invoice"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "delete":{
            "tags":[
               "admin-controller"
            ],
            "summary":"deleteInvoiceById",
            "operationId":"deleteInvoiceByIdUsingDELETE",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{
               "204":{
                  "description":"No Content"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               }
            },
            "deprecated":false
         }
      },
      "/levelUp":{
         "get":{
            "tags":[
               "admin-controller"
            ],
            "summary":"getAllLevelUps",
            "operationId":"getAllLevelUpsUsingGET",
            "produces":[
               "*/*"
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/LevelUp"
                     }
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "post":{
            "tags":[
               "admin-controller"
            ],
            "summary":"createLevelUp",
            "operationId":"createLevelUpUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"levelUp",
                  "description":"levelUp",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/LevelUp"
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Created",
                  "schema":{
                     "$ref":"#/definitions/LevelUp"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "put":{
            "tags":[
               "admin-controller"
            ],
            "summary":"updateLevelUp",
            "operationId":"updateLevelUpUsingPUT_1",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"levelUp",
                  "description":"levelUp",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/LevelUp"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK"
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/levelUp/{id}":{
         "get":{
            "tags":[
               "admin-controller"
            ],
            "summary":"getLevelUpById",
            "operationId":"getLevelUpByIdUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/LevelUp"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "delete":{
            "tags":[
               "admin-controller"
            ],
            "summary":"deleteLevelUpById",
            "operationId":"deleteLevelUpByIdUsingDELETE",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{
               "204":{
                  "description":"No Content"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               }
            },
            "deprecated":false
         }
      },
      "/product":{
         "get":{
            "tags":[
               "admin-controller"
            ],
            "summary":"getAllProducts",
            "operationId":"getAllProductsUsingGET",
            "produces":[
               "*/*"
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/Product"
                     }
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "post":{
            "tags":[
               "admin-controller"
            ],
            "summary":"saveProduct",
            "operationId":"saveProductUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"o",
                  "description":"o",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Product"
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Created",
                  "schema":{
                     "$ref":"#/definitions/Product"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "put":{
            "tags":[
               "admin-controller"
            ],
            "summary":"updateProduct",
            "operationId":"updateProductUsingPUT",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"o",
                  "description":"o",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Product"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK"
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/product/inventory":{
         "put":{
            "tags":[
               "admin-controller"
            ],
            "summary":"updateProductInventory",
            "operationId":"updateProductInventoryUsingPUT",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"inventory",
                  "description":"inventory",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Inventory"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK"
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/product/{id}":{
         "get":{
            "tags":[
               "admin-controller"
            ],
            "summary":"getProduct",
            "operationId":"getProductUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/Product"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "delete":{
            "tags":[
               "admin-controller"
            ],
            "summary":"deleteProduct",
            "operationId":"deleteProductUsingDELETE",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{
               "204":{
                  "description":"No Content"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               }
            },
            "deprecated":false
         }
      }
   },
   "definitions":{
      "Customer":{
         "type":"object",
         "properties":{
            "city":{
               "type":"string"
            },
            "customerId":{
               "type":"integer",
               "format":"int32"
            },
            "email":{
               "type":"string"
            },
            "firstName":{
               "type":"string"
            },
            "lastName":{
               "type":"string"
            },
            "phone":{
               "type":"string"
            },
            "street":{
               "type":"string"
            },
            "zip":{
               "type":"string"
            }
         },
         "title":"Customer"
      },
      "Inventory":{
         "type":"object",
         "properties":{
            "inventory":{
               "type":"integer",
               "format":"int32"
            },
            "productId":{
               "type":"integer",
               "format":"int32"
            }
         },
         "title":"Inventory"
      },
      "Invoice":{
         "type":"object",
         "properties":{
            "customerId":{
               "type":"integer",
               "format":"int32"
            },
            "invoiceId":{
               "type":"integer",
               "format":"int32"
            },
            "purchaseDate":{
               "type":"string",
               "example":"yyyy-MM-dd"
            }
         },
         "title":"Invoice"
      },
      "LevelUp":{
         "type":"object",
         "properties":{
            "customerId":{
               "type":"integer",
               "format":"int32"
            },
            "levelUpId":{
               "type":"integer",
               "format":"int32"
            },
            "memberDate":{
               "type":"string",
               "example":"yyyy-MM-dd"
            },
            "points":{
               "type":"integer",
               "format":"int32"
            }
         },
         "title":"LevelUp"
      },
...
      "Product":{
         "type":"object",
         "properties":{
            "inventory":{
               "type":"integer",
               "format":"int32"
            },
            "list_price":{
               "type":"number",
               "format":"double"
            },
            "productId":{
               "type":"integer",
               "format":"int32"
            },
            "product_description":{
               "type":"string"
            },
            "product_name":{
               "type":"string"
            },
            "unit_price":{
               "type":"number",
               "format":"double"
            }
         },
         "title":"Product"
      },
...
   }
}